Create a CorrGrapheR object

Create a CorrGrapheR object before passing it to plot().

corrgrapher(x, ...)

# S3 method for explainer
corrgrapher(
  x,
  cutoff = 0.2,
  values = NULL,
  cor_functions = list(),
  ...,
  feature_importance = NULL,
  partial_dependence = NULL
)

# S3 method for default
corrgrapher(x, cutoff = 0.2, values = NULL, cor_functions = list(), ...)

Arguments

x
an object to be used to select the method, which must satisfy conditions:

if data.frame (default), columns with type numeric will be selected and called with cor.

if explainer, methods feature_importance and partial_dependence must not return an error. Supply them as arguments(feature_importance or partial_dependence) or supply options to call the functions inside (feature_importance_opts and partial_dependence)

...
other arguments.

cutoff
a number. Correlations below this are treated as no correlation. Edges corresponding to them will not be included in the graph.

values
a data.frame with information about size of the nodes, containing columns value and label (consistent with colnames of x). Default set to equal for all nodes, or (for explainer) importance of variables.

cor_functions
a named list o functions to pass to calculate_cors. Must contain necessary functions from num_num_f, num_cat_f or cat_cat_f. Must contain also max_cor

feature_importance
Either:

an object of feature importance_explainer class, created by feature_importance function, or

a named list of parameters to pass to feature_importance function.

partial_dependence
a named list with 2 elements: numerical and categorical. Both of them should be of aggregated_profile_explainer class, created by partial_dependence function, or a named list of parameters to pass to partial_dependence. If only one kind of data was used, use a list with 1 object.

Value

A corrgrapher object, consisting of following fields:

nodes - a data.frame to pass as argument nodes to visNetwork function

edges - a data.frame to pass as argument edges to visNetwork function

pds (if x was of explainer class) - a splitted aggregated_profiles_explainer object. Each item has information about single variable. Passed to plot.aggregated_profiles_explainer

data - data used to create the object.

See also

Examples

# drop the binary target variable
df <- as.data.frame(datasets::Seatbelts)[,1:7]
cgr <- corrgrapher(df)